+++ lib/generators/thredded/install/templates/initializer.rb
@@ -124,7 +124,7 @@ Thredded.layout = 'thredded/application'
 # Change the HTML sanitization settings used by Thredded.
 # See the Sanitize docs for more information on the underlying library: https://github.com/rgrove/sanitize/#readme
 # E.g. to allow a custom element <custom-element>:
-# Thredded::ContentFormatter.whitelist[:elements] += %w(custom-element)
+# Thredded::ContentFormatter.allowlist[:elements] += %w(custom-element)
 
 # ==> User autocompletion (Private messages and @-mentions)
 # Thredded.autocomplete_min_length = 2 lower to 1 if have 1-letter names -- increase if you want