+++ lib/thredded/content_formatter.rb
@@ -4,8 +4,10 @@ module Thredded
   # Generates HTML from content source.
   class ContentFormatter
     class << self
-      # Sanitization whitelist options.
-      attr_accessor :whitelist
+      # Sanitization allowlist options.
+      attr_accessor :allowlist
+      # TODO: v2.0: drop alias and just use allowlist
+      alias_attribute :whitelist, :allowlist
 
       # Filters that run before processing the markup.
       # input: markup, output: markup.
@@ -28,7 +30,7 @@ module Thredded
       attr_accessor :after_sanitization_filters
     end
 
-    self.whitelist = HTML::Pipeline::SanitizationFilter::WHITELIST.deep_merge(
+    self.allowlist = HTML::Pipeline::SanitizationFilter::WHITELIST.deep_merge(
       elements: HTML::Pipeline::SanitizationFilter::WHITELIST[:elements] + %w[abbr iframe span figure figcaption],
       transformers: HTML::Pipeline::SanitizationFilter::WHITELIST[:transformers] + [
         ->(env) {
@@ -136,7 +138,7 @@ module Thredded
     def content_pipeline_options
       {
         asset_root: Rails.application.config.action_controller.asset_host || '',
-        whitelist: ContentFormatter.whitelist
+        whitelist: ContentFormatter.allowlist
       }
     end
   end